Correct sorting in linking stage of U-Boot build
A problem with the busybox version of sed used in Baserock means
that the required \a bell character is not properly escaped in
sed substitutions, causing sorting to go wrong, and leading to
a built U-Boot which has no commands. This has been refactored
in newer U-Boot versions, so should no longer be a concern.

Here, due to the complexity of building and debugging a mainline
U-Boot SPL preloader which works on this board, we have used this
old version of U-Boot, as time is not available to do the work
to get this to mainline.